refactor(voice-call): share tts deep merge
(cherry picked from commit ed437434afcdb5f2819f6e1f47b6c88cb7e8bf6f)
